Support large tile values in achievement snapshots
Changed GameSnapshot.max_tile from Int to Double to handle tile values beyond Int.max. Updated AchievementEvaluator to compute tile values from step using Double, and adjusted AchievementStore to use and log Double values for tile progression achievements. Added debug print statements for better traceability and improved error handling when loading persisted snapshots.

+++ Packages/GameServices/Sources/GameServices/AchievementModels.swift
@@ -162,7 +162,7 @@ public struct GameSnapshot: Sendable, Codable {
     public var board_monotonic_end: Bool = false
     public var made_2244_square: Bool = false
     
-    public var max_tile: Int = 0
+    public var max_tile: Double = 0
     public var win: Bool = false
     public var total_moves: Int = 0
     public var combo610Total: Int = 0
